accessware/backend/requirements.txt
fastapi
uvicorn[standard]
pyserial
pytest
pytest-asyncio
httpx
